requirements.txt
PyYAML
pyinstaller
deepl
keyboard
pystray